♻️ [#87] Refactor: Button 리팩토링
