+++ src/widgets/product-detail/components/product-info/index.ts
@@ -1,6 +1,3 @@
 export { default as ProductInfoWrapper } from './wrapper/ProductInfoWrapper'
 export { default as ProductContents } from './product-contents/ProductContents'
 export { default as ProductImage } from './product-image/ProductImage'
-export { default as LikeButton } from './like-button/LikeButton'
-export { default as ClipBoardButton } from './clip-board-button/ClipBoardButton'
-export { default as KaKaoShareButton } from './kakao-share-button/KaKaoShareButton'
