+++ src/widgets/product-detail/components/product-info/product-contents/ClipBoardButton.tsx
